requirements.txt
streamlit==1.32.0
pillow==10.1.0
opencv-python==4.9.0.80
numpy==1.26.3
insightface==0.7.3
onnxruntime==1.16.3
scikit-learn==1.3.2
